Use common function to add message to ILP encoder
